+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+### Fixed
+
+- Fixed crash when exiting the app prematurely https://github.com/Textualize/textual/pull/5039
+
 ## [0.80.0] - 2024-09-23
 
 ### Added