+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.80.1] - 2024-09-24
 
-### [0.80.1] - 2024-09-24
+### Fixed
 
 - Fixed crash when exiting the app prematurely https://github.com/Textualize/textual/pull/5039
 - Fixed exception constructing TextArea outside of App https://github.com/Textualize/textual/pull/5045